requirements.txt
openai==1.12.0
httpx==0.27.0
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pyyaml==6.0.1
pydantic==2.6.1
slowapi==0.1.8
tenacity==8.2.3